--- a/clojure/project.clj
+++ b/clojure/project.clj
@@ -6,6 +6,8 @@
   :dependencies [[org.clojure/clojure "1.11.1"]
                  ; [cheshire "5.12.0"]
                  ; [metosin/jsonista "0.3.8"]
+                 ; [com.dslplatform/dsl-json "2.0.2"]
+                 ; [com.alibaba/fastjson "2.0.41"]
                  [com.google.code.gson/gson "2.10.1"]]
 
   :profiles {:uberjar {:aot :all} }