@jinyus 2023-10-17
This Pr:

```
Clojure:

        Benchmark 1: java  -jar ./target/related.jar
        Processing time (w/o IO): 4088ms
        Processing time (w/o IO): 3589ms
        Processing time (w/o IO): 4108ms
        Processing time (w/o IO): 3610ms
        Processing time (w/o IO): 3631ms
          Time (mean ± σ):      4.651 s ±  0.215 s    [User: 6.268 s, System: 0.162 s]
          Range (min … max):    4.473 s …  5.068 s    10 runs
         
Clojure:

        Benchmark 1: java  -jar ./target/related.jar
        Processing time (w/o IO): 50047ms
        Processing time (w/o IO): 51475ms
        Processing time (w/o IO): 58079ms
          Time (mean ± σ):     55.984 s ±  4.615 s    [User: 58.385 s, System: 0.264 s]
          Range (min … max):   52.721 s … 59.248 s    2 runs
         
Clojure:

        Benchmark 1: java  -jar ./target/related.jar
        Processing time (w/o IO): 506421ms
        Processing time (w/o IO): 439515ms
        Processing time (w/o IO): 439554ms
          Time (mean ± σ):     441.289 s ±  0.071 s    [User: 444.729 s, System: 0.592 s]
          Range (min … max):   441.239 s … 441.340 s    2 runs
```

Main Branch:

```
Clojure:

        Benchmark 1: java  -jar ./target/related.jar
        Processing time (w/o IO): 3643ms
        Processing time (w/o IO): 4203ms
        Processing time (w/o IO): 4082ms
        Processing time (w/o IO): 4035ms
        Processing time (w/o IO): 4063ms
        Processing time (w/o IO): 3665ms
          Time (mean ± σ):      4.949 s ±  0.238 s    [User: 7.107 s, System: 0.156 s]
          Range (min … max):    4.581 s …  5.184 s    10 runs
         
Clojure:

        Benchmark 1: java  -jar ./target/related.jar
        Processing time (w/o IO): 48831ms
        Processing time (w/o IO): 54383ms
        Processing time (w/o IO): 48658ms
          Time (mean ± σ):     52.742 s ±  4.049 s    [User: 55.084 s, System: 0.188 s]
          Range (min … max):   49.879 s … 55.605 s    2 runs
         
Clojure:

        Benchmark 1: java  -jar ./target/related.jar
        Processing time (w/o IO): 431250ms
        Processing time (w/o IO): 427395ms
        Processing time (w/o IO): 427454ms
          Time (mean ± σ):     429.242 s ±  0.029 s    [User: 431.881 s, System: 0.460 s]
          Range (min … max):   429.222 s … 429.262 s    2 runs

```